fix: acosf -> acos

+++ srcs/raytrace/calc_specular_reflection.c
@@ -19,7 +19,7 @@ bool	SPOT_check(t_vec *dir_pos2lgt_n, t_light *lgt_inf)
 	double	alpha;
 
 	inverse_vec(&pos2lgt, dir_pos2lgt_n);
-	alpha = acosf(dot_vec(&pos2lgt, &lgt_inf->direction));
+	alpha = acos(dot_vec(&pos2lgt, &lgt_inf->direction));
 	if (alpha > lgt_inf->sl_angle / 2.0f * (double)M_PI / 180.0)
 		return (false);
 	return (true);
